test: cleanups
